Add macros to log duration measurements using a timing event

+++ c8/stm/prof.c
@@ -42,8 +42,8 @@ static void _stm_profiling_event(stm_thread_local_t *tl,
             struct timespec *duration = payload->data.duration;
             buf.extra_length = sprintf(buf.extra,
                                         "s%un%u",
-                                        duration->tv_sec,
-                                        duration->tv_nsec);
+                                        (uint32_t)duration->tv_sec,
+                                        (uint32_t)duration->tv_nsec);
         }
     }
